refactor: moved filter interfaces to types.ts and enum to enums.ts
